Added variables and added also variable evaluation using expressions. Variables are stored in session and are retained until overwritten or a new session is used.

+++ src/Systemic/QueryResult/QueryResult.cpp
@@ -23,7 +23,7 @@ void QueryResult::Print() const{
 
       switch (column.GetType()){
         case DataType::TinyInt:
-          std::cout << column.GetTinyInt();
+          std::cout << static_cast<int16_t>(column.GetTinyInt());
           break;
         case DataType::SmallInt:
           std::cout << column.GetSmallInt();
